created DiveLogView with cell contents, added view to TableViewCell with constraints, changed timestamp to 3 letter day
